Moved Chess into examples
